+++ corp104/cmd/server/handler.go
@@ -28,6 +28,7 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/goincremental/negroni-sessions"
 	"github.com/gorilla/context"
 	"github.com/julienschmidt/httprouter"
 	"github.com/meatballhat/negroni-logrus"
@@ -56,6 +57,7 @@ func EnhanceRouter(c *config.Config, cmd *cobra.Command, serverHandler *Handler,
 	for _, m := range middlewares {
 		n.Use(m)
 	}
+	n.Use(sessions.Sessions(c.GetWebSessionName(), c.Context().WebSession.Store))
 	n.UseFunc(serverHandler.RejectInsecureRequests)
 	n.UseHandler(router)
 	if enableCors {
