+++ corp104/config/context.go
@@ -36,4 +36,5 @@ type Context struct {
 	FositeStore    pkg.FositeStorer
 	KeyManager     jwk.Manager
 	ConsentManager consent.Manager
+	WebSession     *WebSession
 }
